all but last

+++ 11.Crazy.Tower/Hero.hpp
@@ -14,20 +14,20 @@ struct Hero;
 
 struct RoomOption {
     pStairs stairs;
-    std::size_t multiplicity;
-    RoomOption(pStairs stairs, const std::size_t& multiplicity) : stairs(stairs), multiplicity(multiplicity) {};
-    virtual std::size_t exec(Hero& hero, const std::size_t& init_stamina);
+    std::uint32_t multiplicity;
+    RoomOption(pStairs stairs, const std::uint32_t& multiplicity) : stairs(stairs), multiplicity(multiplicity) {};
+    virtual std::uint32_t exec(Hero& hero, const std::size_t& init_stamina);
     };
 
 
 struct Hero {
     const std::size_t max_stamina;
     std::size_t stamina;
-    const std::size_t modulo;
+    const std::uint32_t modulo;
 
-    Hero(const std::size_t& max_stamina, const std::size_t& stamina, const std::size_t& modulo) : max_stamina(max_stamina), stamina(stamina), modulo(modulo) {};
+    Hero(const std::size_t& max_stamina, const std::size_t& stamina, const std::uint32_t& modulo) : max_stamina(max_stamina), stamina(stamina), modulo(modulo) {};
 
-    static std::size_t visit(const pRoom& room, Hero& hero) {
+    static std::uint32_t visit(const pRoom& room, Hero& hero) {
         DEBUGGING(std::cout << "At room '" << room->id << "' with [" << room->downstairs.size() << "] stairs" << std::endl)
         DEBUGGING(std::cout << "\t stamina: " << hero.stamina << std::endl)
         if (room->is_exit) {
@@ -38,7 +38,7 @@ struct Hero {
         auto stamina = hero.stamina; // 'hero' has this level of stamina on arrival
         auto options = Hero::generate(room, hero);
         DEBUGGING(std::cout << "\t>> " << options.size() << " options" << std::endl)
-        std::size_t successful_options = 0;
+        std::uint32_t successful_options = 0;
         for (auto it = options.begin(); it!=options.end(); ++it) {
             auto new_solutions = it->exec(hero, stamina);
             successful_options = ModularArithmetic<std::uint32_t>::sum(successful_options, new_solutions, hero.modulo);
@@ -62,9 +62,9 @@ struct Hero {
             possible &= (needed <= static_cast<int>(room->downstairs.size()));
             DEBUGGING(if (!possible) {std::cout << "\t\t not enough minions to get '" << needed << "' STAMINA!" << std::endl;})
             if (possible) {
-                auto multiplicity = ModularArithmetic<std::uint32_t>::combinations(room->downstairs.size()-1, std::max<int>(0, (*it)->keys-1), hero.modulo);
+                auto multiplicity = ModularArithmetic<std::uint32_t>::combinations(room->downstairs.size()-1, std::max<int>(0, static_cast<int>((*it)->keys)-1), hero.modulo);
                 options.insert(options.end(), RoomOption(*it, multiplicity));
-                DEBUGGING(std::cout << "\t\t >> " << multiplicity << " ways (C_{" << room->downstairs.size()-1 << ", " << std::max<int>(0, (*it)->keys-1) << "})" << std::endl)
+                DEBUGGING(std::cout << "\t\t >> " << multiplicity << " ways (C_{" << room->downstairs.size()-1 << ", " << std::max<int>(0, static_cast<int>((*it)->keys)-1) << "})" << std::endl)
                 }
         
             }
@@ -75,7 +75,7 @@ struct Hero {
 
 
 /* RoomOption implementation */
-std::size_t RoomOption::exec(Hero& hero, const std::size_t& init_stamina) {
+std::uint32_t RoomOption::exec(Hero& hero, const std::size_t& init_stamina) {
     // We have to kill a number of minions equal to the number of keys (and then go down through the stairs)
     hero.stamina = std::min(hero.max_stamina, init_stamina + std::max(1, stairs->keys)); // There is always a minion to kill :'(
     hero.stamina = std::min(hero.max_stamina, hero.stamina - stairs->stamina);
